Player movement.

+++ include/player_config.h
@@ -1,10 +1,10 @@
 #pragma once
 
 namespace playerConfig {
-    extern int MAX_DX;
-    extern int MAX_DY;
-    extern int MAX_DDX;
-    extern int MAX_DDY;
+    extern double MAX_DX;
+    extern double MAX_DY;
+    extern double MAX_DDX;
+    extern double MAX_DDY;
     extern int INITIAL_X; 
     extern int INITIAL_Y;
 }
